Minor updat

+++ firmware_v3/datalogger/datalogger.ino
@@ -175,9 +175,9 @@ public:
                 switch (index) {
                 case 0:
                     if (gd.date != v) {
-                      byte year = v % 100;
+                      int year = v % 100;
                       // filter out invalid date
-                      if (v < 1000000 && v >= 10000 && year >= 15 && (gd.date == 0 || (int)year - (gd.date % 100) <= 1)) {
+                      if (v < 1000000 && v >= 10000 && year >= 15 && (gd.date == 0 || year - (gd.date % 100) <= 1)) {
                         gd.date = v;
                         mask |= 0x1;
                       }
